update and move the directory to partial test the application
