+++ src/lib/open.js
@@ -7,8 +7,8 @@ const open = require('opn');
  * @return {boolean} true on open false on failed
  */
 module.exports = function(config = {}) {
-  if (config.open === false) {
-    return false;
+  if (config.open === false || process.env.NODE_ENV === 'test') {
+    return true;
   }
   let args = [];
   // If there is just the true option then we need to construct the link
